+++ packages/orq-rc/src/orq_ai_sdk/models/conversationresponse.py
@@ -104,7 +104,7 @@ class ConversationResponse(BaseModel):
     r"""Unix timestamp in milliseconds when the conversation was last modified."""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "conv_01khts5np47emn01b8vpyfzkqp"
+        "conv_01khwc0xkfx3g2d9ahsprsz9e2"
     )
     r"""Unique conversation identifier with `conv_` prefix."""