+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -314,7 +314,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KHTS5PY2KEX1QPG53A9T7FFQ"
+        "01KHWC0Y3KF9EBFY8YG6XYHJ5J"
     )
     r"""The unique identifier of the data source"""