+++ packages/orq-rc/src/orq_ai_sdk/models/updatetoolop.py
@@ -259,7 +259,7 @@ class RequestBodyTools(BaseModel):
 
     schema_: Annotated[UpdateToolRequestBodyToolsSchema, pydantic.Field(alias="schema")]
 
-    id: Optional[str] = "01KHTS5PSPKY00E7S1TPAKCA47"
+    id: Optional[str] = "01KHWC0XZ39XRVHNKREXJFCW87"
 
     description: Optional[str] = None
 
@@ -1214,7 +1214,7 @@ class UpdateToolResponseBodyCodeExecutionTool(BaseModel):
     code_tool: UpdateToolResponseBodyCodeTool
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "tool_01KHTS5PSGF45ETBEHMZP3AY5N"
+        "tool_01KHWC0XYXSKHS1KWYVSP3YNVS"
     )
 
     display_name: Optional[str] = None
@@ -1345,7 +1345,7 @@ class UpdateToolResponseBodyTools(BaseModel):
         UpdateToolResponseBodyToolsSchema, pydantic.Field(alias="schema")
     ]
 
-    id: Optional[str] = "01KHTS5PSF98S93S17A9MBE0N7"
+    id: Optional[str] = "01KHWC0XYWMARG8W33B57S7QQB"
 
     description: Optional[str] = None
 
@@ -1476,7 +1476,7 @@ class UpdateToolResponseBodyMCPTool(BaseModel):
     mcp: UpdateToolResponseBodyMcp
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "tool_01KHTS5PS6E5465KMBXHS9XPWH"
+        "tool_01KHWC0XYV2N3C78854EN8EMKD"
     )
 
     display_name: Optional[str] = None
@@ -1777,7 +1777,7 @@ class UpdateToolResponseBodyHTTPTool(BaseModel):
     http: UpdateToolResponseBodyHTTP
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "tool_01KHTS5PS5SPJM9G383WV2CYYQ"
+        "tool_01KHWC0XYS2EYSW6TJAWQC9WAN"
     )
 
     display_name: Optional[str] = None
@@ -1972,7 +1972,7 @@ class UpdateToolResponseBodyJSONSchemaTool(BaseModel):
     json_schema: UpdateToolResponseBodyJSONSchema
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "tool_01KHTS5PS3BJ5993DRK8V394QB"
+        "tool_01KHWC0XYRRSHZ8ANWZ8QR94BB"
     )
 
     display_name: Optional[str] = None
@@ -2171,7 +2171,7 @@ class UpdateToolResponseBodyFunctionTool(BaseModel):
     function: UpdateToolResponseBodyFunction
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "tool_01KHTS5PS26S752W8XCBGQMS52"
+        "tool_01KHWC0XYQNZM4WR71FHKN5EQ0"
     )
 
     display_name: Optional[str] = None